Progress towards refactoring into an Internal module to allow the FixedPoint module to provide alternative arithmetic operations without exposing internals.
